#10 코드리팩토링 - 기능 분리, 라우터클래스, 테스트코드이동
