+++ bin/www.js
@@ -1,0 +1,5 @@
+const  app = require('../index');
+
+app.listen(3000, () => {
+    console.log('Server is running on 3000 port!');
+})
