+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "basic-auth": "^2.0.1",
     "body-parser": "^1.19.0",
+    "cross-env": "^7.0.2",
     "express": "^4.17.1",
     "morgan": "^1.10.0"
   },
@@ -15,8 +16,8 @@
     "supertest": "^4.0.2"
   },
   "scripts": {
-    "test": "mocha index.spec.js",
-    "start": "node index.js"
+    "test": "cross-env NODE_ENV=test mocha api/user/user.spec.js",
+    "start": "node bin/www.js"
   },
   "repository": {
     "type": "git",
